+++ src/Response/V1/Address.php
@@ -192,6 +192,14 @@ class Address implements JsonSerializable
     }
 
     /**
+     * @return array<string,mixed>
+     */
+    public function __debugInfo()
+    {
+        return $this->jsonSerialize();
+    }
+
+    /**
      * @return array{
      *     jisx0402: string,
      *     old_code: string,